# docs/admin/admin_swagger.yaml
basePath: /api
definitions:
...
  admin.ChangeCurPasswordForm:
    properties:
      new_password:
        maxLength: 20
        minLength: 4
        type: string
      old_password:
        maxLength: 20
        minLength: 4
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  admin.LoginPayload:
    properties:
      nickname:
        type: string
      route_names:
        items:
          type: string
        type: array
      token:
        type: string
      username:
        type: string
    type: object
  admin.PeerForm:
    properties:
      cpu:
        type: string
      hostname:
        type: string
      id:
        type: string
      memory:
        type: string
      os:
        type: string
      row_id:
        type: integer
      username:
        type: string
      uuid:
        type: string
      version:
        type: string
    type: object
...
  admin.UserForm:
    properties:
      avatar:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      is_admin:
        type: boolean
      nickname:
        description: Password string           `json:"password" validate:"required,gte=4,lte=20"`
        type: string
      status:
        allOf:
        - $ref: '#/definitions/model.StatusCode'
        minimum: 0
      username:
        maxLength: 10
        minLength: 4
        type: string
    required:
    - group_id
    - nickname
    - status
    - username
    type: object
  admin.UserPasswordForm:
    properties:
      id:
        type: integer
      password:
        maxLength: 20
        minLength: 4
        type: string
    required:
    - id
    - password
    type: object
...
  model.Peer:
    properties:
      cpu:
        type: string
      created_at:
        type: string
      hostname:
        type: string
      id:
        type: string
      memory:
        type: string
      os:
        type: string
      row_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
      username:
        type: string
      uuid:
        type: string
      version:
        type: string
    type: object
  model.PeerList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Peer'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  model.StatusCode:
    enum:
    - 1
    - 2
    type: integer
    x-enum-comments:
      COMMON_STATUS_DISABLED: 通用状态 禁用
      COMMON_STATUS_ENABLE: 通用状态 启用
    x-enum-varnames:
    - COMMON_STATUS_ENABLE
    - COMMON_STATUS_DISABLED
...
  model.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      is_admin:
        type: boolean
      nickname:
        type: string
      status:
        $ref: '#/definitions/model.StatusCode'
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.UserList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.User'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: 接口
  title: 管理系统API
  version: "1.0"
paths:
...
  /admin/peer/create:
    post:
      consumes:
      - application/json
      description: 创建机器
      parameters:
      - description: 机器信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.PeerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Peer'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 创建机器
      tags:
      - 机器
  /admin/peer/delete:
    post:
      consumes:
      - application/json
      description: 机器删除
      parameters:
      - description: 机器信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.PeerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 机器删除
      tags:
      - 机器
  /admin/peer/detail/{id}:
    get:
      consumes:
      - application/json
      description: 机器详情
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Peer'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 机器详情
      tags:
      - 机器
  /admin/peer/list:
    get:
      consumes:
      - application/json
      description: 机器列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.PeerList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 机器列表
      tags:
      - 机器
  /admin/peer/update:
    post:
      consumes:
      - application/json
      description: 机器编辑
      parameters:
      - description: 机器信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.PeerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Peer'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 机器编辑
      tags:
      - 机器
...
  /admin/user/changeCurPwd:
    post:
      consumes:
      - application/json
      description: 修改当前用户密码
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.ChangeCurPasswordForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 修改当前用户密码
      tags:
      - 用户
  /admin/user/create:
    post:
      consumes:
      - application/json
      description: 创建管理员
      parameters:
      - description: 管理员信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UserForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 创建管理员
      tags:
      - 用户
  /admin/user/current:
    get:
      consumes:
      - application/json
      description: 当前用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.LoginPayload'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 当前用户
      tags:
      - 用户
  /admin/user/delete:
    post:
      consumes:
      - application/json
      description: 管理员编删除
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UserForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员删除
      tags:
      - 用户
  /admin/user/detail/{id}:
    get:
      consumes:
      - application/json
      description: 管理员详情
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员详情
      tags:
      - 用户
  /admin/user/list:
    get:
      consumes:
      - application/json
      description: 管理员列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: page_size
        type: integer
      - description: 账户
        in: query
        name: username
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员列表
      tags:
      - 用户
  /admin/user/update:
    post:
      consumes:
      - application/json
      description: 管理员编辑
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UserForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 管理员编辑
      tags:
      - 用户
  /admin/user/updatePassword:
    post:
      consumes:
      - application/json
      description: 修改密码
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UserPasswordForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: 修改密码
      tags:
      - 用户
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"